refactor to use Dash Pages for main content
See https://dash.plotly.com/urls
